requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.11
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-localflavor==3.1
Pillow==9.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
